Improvement
Add some improvement in the test about the logger. Create a project test starter.
