+++ Learning_test_automation/Spotkanie8_Zadania/Hotel/Logger/NLog.config
@@ -15,8 +15,7 @@
     <!-- Log Error and above level messages to the console -->
     <logger name="*" minlevel="Error" writeTo="console" />
     <!-- Log Warn and above level messages to the console -->
-    <logger name="*" minlevel="Warn" writeTo="console" />
-    
+    <logger name="*" minlevel="Warn" writeTo="console" />  
   </rules>
 </nlog>
