+++ Learning_test_automation/Spotkanie8_Zadania/Hotel/TestStarter/TestStarter.csproj
@@ -1,0 +1,18 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Selenium.WebDriver.ChromeDriver" Version="116.0.5845.9600" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\WrapSeleniumMethod\SeleniumDriver.csproj" />
+  </ItemGroup>
+
+</Project>
