+++ Learning_test_automation/Spotkanie8_Zadania/Hotel/WrapSeleniumMethod/ISeleniumDriver.cs
@@ -17,6 +17,10 @@ namespace SeleniumDriver
 
         void ComparingTexts(IWebElement lookForItem, string searchText);
 
-        void WaitFor(int amount); 
+        void WaitFor(int amount);
+
+        IWebElement DayOfMonthSelection(IWebElement packingForNumberOfDayAndAvailability, int numberDayOfMonth);
+
+        void MarkingDateOfBookingRoom(IWebElement firstDay, IWebElement lastDay);
     }
 }
