+++ Learning_test_automation/Spotkanie8_Zadania/Vitodens343-F_TestsConfig.json
@@ -1,0 +1,12 @@
+{
+    "Name": "Vitodens343-F",
+    "Family": "Vitodens3xx",
+    "LeafSerialNumber": "92908",
+    "WifiConnectionSsid": "Viessmann-00124183",
+    "WifiConnectionTischmodelPassword": "00124183",
+    "Vin": "7722304000000003",
+    "HmuSoftwareVersion": "2.514.2008.423",
+    "BcuSoftwareVersion": "1.519.2004.136",
+    "HmiSoftwareVersion": "3.524.2004.1",
+    "TcuSoftwareVersion": "19.501.2044.85"
+}
